refactor: split paasng.platform.applications.views
